Move isValidFilePath from FilePathUtil into AddressBook. Fix comments to not specify operating system's specific reserved characters.
